requirements.txt
alembic==1.13.2
aniso8601==9.0.1
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
email-validator==2.2.0
et-xmlfile==1.1.0
flasgger==0.9.7.1
flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
flask-sqlalchemy==3.1.1
greenlet==3.0.3
idna==3.8
importlib-metadata==8.2.0
importlib-resources==6.4.0
itsdangerous==2.2.0
jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
MarkupSafe==2.1.5
mistune==3.0.2
mysql-connector-python==9.0.0
mysqlclient==2.2.4
numpy==1.24.4
openpyxl==3.1.5
packaging==24.1
pandas==2.0.3
pkgutil-resolve-name==1.3.10
py3dns==4.0.2
PyJWT==2.8.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
SQLAlchemy==2.0.31
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
validate-email-address==1
werkzeug==3.0.3
zipp==3.19.2
